fix(slider-web): finish with loading
