--- packages/pluggableWidgets/slider-web/src/utils/isValidMinMax.ts
@@ -1,9 +1,0 @@
-import { SliderContainerProps } from "../../typings/SliderProps";
-import { getMinValue } from "./getMinValue";
-import { getMaxValue } from "./getMaxValue";
-
-export function isValidMinMax(props: SliderContainerProps): boolean {
-    const min = getMinValue(props);
-    const max = getMaxValue(props);
-    return min < max;
-}
