--- packages/pluggableWidgets/slider-web/src/utils/isVertical.ts
@@ -1,5 +1,0 @@
-import { SliderContainerProps } from "../../typings/SliderProps";
-
-export function isVertical(params: Pick<SliderContainerProps, "orientation">): boolean {
-    return params.orientation === "vertical";
-}
